Worked on better management of the errors in the API

+++ DigiLog/DigiLog/DTOs/EmployeeDTO.cs
@@ -2,7 +2,7 @@
 {
     public class EmployeeDTO
     {
-        internal int Id { get; set; }
+        internal long Id { get; set; }
         public string? FirstName { get; set; }
         public string? MiddleName { get; set; }
         public string? LastName { get; set; }
